Still working on lee.py process_alternate shape fixes
I think the problem is in the fix_dims_quad function. This
doesn't have the right logic when the JointDistribution from
which the 'quad' function was generated (and from where the
batch_shape is currently derived) has 1 event and 1 hypothesis
for batch_shape of (). The number of events used to get the
actual quad function from this doesn't factor into the batch
shape. So that could be an alternate fix; make the batch shape
properly aware of the number of events "built in" to the quad
function.